+++ sdk/eventhub/azure-eventhub/azure/eventhub/_client_base.py
@@ -341,6 +341,7 @@ class ClientBase(object):  # pylint:disable=too-many-instance-attributes
             custom_endpoint_hostname=self._config.custom_endpoint_hostname,
             port=self._config.connection_port,
             verify=self._config.connection_verify,
+            refresh_window=300,
         )
 
     def _close_connection(self):